refactor: move picker test
